+++ codificacao/src/main/java/app/controllers/RoadController.java
@@ -41,8 +41,18 @@ public class RoadController extends GenericAppController {
         try{
             Base.open(Db.getInstance());
             Road road = Road.findById(Integer.parseInt(resourceId));
-            RoadJson stateJson = new RoadJson(road);
-            ctx.result(mapper.writeValueAsString(stateJson));
+            if(road == null){
+                ctx.res.setStatus(404);
+            }
+            else {
+                if (road.delete()) {
+                    ctx.res.setStatus(200);
+                    RoadJson roadJson = new RoadJson(road);
+                    ctx.result(mapper.writeValueAsString(roadJson));
+                } else {
+                    ctx.res.setStatus(400);
+                }
+            }
             Base.close();
         }
         catch (Exception e){
@@ -101,11 +111,15 @@ public class RoadController extends GenericAppController {
         try{
             Base.open(Db.getInstance());
             Road road = Road.findById(Integer.parseInt(resourceId));
-            if(road.delete()){
-                ctx.res.setStatus(200);
+            if(road == null){
+                ctx.res.setStatus(404);
             }
-            else{
-                ctx.res.setStatus(400);
+            else {
+                if (road.delete()) {
+                    ctx.res.setStatus(200);
+                } else {
+                    ctx.res.setStatus(400);
+                }
             }
             Base.close();
         }
